Refactored a math optimizers
